fix: regression bug-fix
1. Add redirection in dashboard.index route
