+++ app/dashboard/index/template.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
