+++ app/dashboard/show/route.js
@@ -5,6 +5,7 @@ export default Route.extend(AuthenticatedRouteMixin, {
   model(params) {
     return this.store.findRecord('collection', params.collection_id);
   },
+
   actions: {
     error() {
       return this.transitionTo('/404');
